Revert previous change and clairyfy usage
- A redirect rule still needs a lower priority pass rule for the same version.
While it may seem redundant, there's no guarantee that the redirect target
is on the same tier or running under the same config domain, so there will
have to be a pass rule there, no matter what. Thus special casing for redirects
on the same tier would not help usability.