Temporary bypass of account creation flag to fix user creation problem
One existing client doesn't pass True for automatic_account_creation which is
required for new users to be able to register. This has been fixed in all other clients
and the next version of the specific client. Force account creation for the affected
auth provider types until new clients can be deployed.

Clients requiring the proper authentication flow for new identities are not using
these specific providers at the moment.