Make 'ami run' deploy new instances immediately
The 'ami run' command will terminate all instances in the autoscaling group so that new ones will be launched with the new AMI. The instances have lifecycle hook which means they will linger on for two minutes while all pending connections are drained and new instances are being launched.

The command will no longer spew out acres of text by default. Use the --verbose switch to get the acres back.

NOTE! If The launch configuration is associated with a classic load balancer, the instances will be removed from the ELB immediately. This is quite unfortunate, especially for the API router app. Updating the 'ami' command to use targets groups may fix this.