Fix remote address detection in lambda apps
The X-Forwarded-For item in the request header is set by nginx api router but is overwritten by AWS API Gateway because of reasons. The api router will instead use X-Real-IP to pass the proper remote address onwards.
NOTE! The wsgi handler must be certain that the X-Real-IP was indeed set by the api router but it's hard to do so reliably but if the callers ip starts with "10." then it's assumed it's running from within the VPC. Not great but should work.

The root endpoint info contains some more info in "debug" mode. If caller has service role the extra info is also displayed.