Fix authentication provider for systems tests
The authentication endpoint was only working for drift-base. This was due to some fallout after authentication logic refactoring. The systesthelper module in drift still needs to detect if 'drift-base' is running or some othe project but that can be fixed by moving the code where it switches providers into drift-base itself.

Also, the 'auth' endpoint added to root doc.