Fix how current tenant name is discovered
The 'tenant_from_hostname' property was assuming a Flask request context which is not always the case.
Renaming the property to 'tenant_name' and deprecating the old name. The tenant name is not always derived from host name. It can also come from a request header or from an environment variable.