Integrate alembic into Drift
The env.py module should only contain the following lines:
    from drift.contrib.alembic import run
    run()

The db model of the current project is made available to alembic which means that the --autogenerate switch works.

To prep a project for alembic run 'alembic init' and update the env.py file as described above.

To create an update:
alembic revision --autogenerate -m "Some text description"