[postgres] Add 'pre_pool_ping=True' to the engine config
- If the DB is restarted or for some other reason drops the connection,
all subsequent connections would fail until the pool had been emptied.
Pre-ping will pass a minimal command to the server to verify the
connection before it's given to the API handler, ensuring that a new
connection is opened if the old one is broken.
- This doesn't handle if the connection goes down mid-transaction, but
that's a larger project to handle.