[redis] Hard code decode_responses to True
- It needs to otherwise be configured for every tenant in a deployable
which is error prone, and drift itself would have to check the setting
for anything using Redis. The flag was there to be backwards compatible
with Python 2 which we don't care about anymore, so we can hard code it
to work the way we need it to. If we ever need non-decoded responses, we
can make a custom connection on the side with different settings and use
that explicitly.